Move ALLOW_CLICK logic into `SlidedownManager`
Motivation: since the allow click refers specifically to the allow click on the slidedown, it makes more sense to move it to its dedicated class. It also makes the `PromptsManager` observers cleaner.
