+++ src/managers/slidedownManager/types.ts
@@ -4,4 +4,5 @@ export interface ISlidedownManager {
     setIsSlidedownShowing: (isShowing: boolean) => void;
     showQueued: () => Promise<void>;
     createSlidedown: (options: AutoPromptOptions) => Promise<void>;
+    handleAllowClick: () => Promise<void>;
 }
